+++ app/src/main/java/com/example/app/common/listener/OnItemEventListener.java
@@ -3,9 +3,9 @@ package com.example.app.common.listener;
 import android.os.Bundle;
 
 public interface OnItemEventListener {
-    default void onItemEvent(int sortedIndex, Bundle data) {
+    default void onItemEvent(long sortedIndex, Bundle data) {
         onItemEvent(sortedIndex, false, data);
     }
 
-    void onItemEvent(int sortedIndex, boolean switchValue, Bundle data);
+    void onItemEvent(long sortedIndex, boolean switchValue, Bundle data);
 }
